Fix for issues #28 and #29. (#30)
For issue and bug #28, the GNU and Intel debug compilations are not
apples-to-apples.  The GNU debug compilation is at -O1 and the Intel
debug compilation is at -O0.  The compiler creates code motion at
higher and higher optimizaiton levels.  Code motion means that the
assembly instructions are moved around so the CPU can execute the
instructions faster.  To help debug a program, it is helpful to
group the assembly instructions with their source code lines.  Code
motion ungroups the instructions, which makes it harder to debug.
So -O0 is the best optimization for debug.  So we are setting the
GNU debug level to -O0.  This makes it apples-to-apples with the
Intel debug level and helps with debugging.

For issue and bug #29, the GNU optimization is -O2 at the release level.
Unfortunately, the GNU compiler does not vectorize at -O2.  So we
are changing the optimization level to -O3, which turns on
vectorization.  This makes it apples-to-apples with the Intel
release optimization level.

Co-authored-by: Christopher Harrop <35781497+christopherwharrop-noaa@users.noreply.github.com>